webapi: Extract handlers to webapi, move api to trcapi
