+++ cmd/srrs/main.go
@@ -1,10 +1,7 @@
 package main
 
 import (
-	"compress/flate"
-	"encoding/json"
 	"flag"
-	"fmt"
 	"log"
 	"net"
 	"net/http"
@@ -14,9 +11,9 @@ import (
 	"strings"
 	"time"
 
-	"github.com/gorilla/websocket"
 	"github.com/pkg/errors"
-	"github.com/rvolosatovs/turtlitto/pkg/api"
+	"github.com/rvolosatovs/turtlitto/pkg/trcapi"
+	"github.com/rvolosatovs/turtlitto/pkg/webapi"
 )
 
 const (
@@ -28,12 +25,6 @@ var (
 	httpAddr = flag.String("http", defaultAddr, "HTTP service address")
 	static   = flag.String("static", "", "Path to the static assets")
 	sock     = flag.String("socket", filepath.Join(os.TempDir(), "trc.sock"), "Path to the unix socket")
-	upgrader = websocket.Upgrader{
-		EnableCompression: true,
-		CheckOrigin: func(r *http.Request) bool {
-			return true
-		},
-	}
 
 	statusEndpoint  = path.Join("api", "v1", "status")
 	turtleEndpoint  = path.Join("api", "v1", "turtles")
@@ -47,13 +38,13 @@ func init() {
 func main() {
 	flag.Parse()
 
-	pool := api.NewPool(func() (*api.Conn, func(), error) {
+	pool := trcapi.NewPool(func() (*trcapi.Conn, func(), error) {
 		unixConn, err := net.Dial("unix", *sock)
 		if err != nil {
 			return nil, nil, errors.Wrapf(err, "failed to connect to TRC's unix socket")
 		}
 
-		trcConn, err := api.Connect(api.DefaultVersion, unixConn, unixConn)
+		trcConn, err := trcapi.Connect(trcapi.DefaultVersion, unixConn, unixConn)
 		if err != nil {
 			return nil, nil, errors.Wrapf(err, "failed to establish connection to TRC")
 		}
@@ -87,168 +78,12 @@ func main() {
 		}
 	}()
 
-	http.HandleFunc("/"+statusEndpoint, func(w http.ResponseWriter, r *http.Request) {
-		defer r.Body.Close()
-
-		// TODO: Check token
-
-		wsConn, err := upgrader.Upgrade(w, r, nil)
-		if err != nil {
-			http.Error(w, fmt.Sprintf("Failed to open WebSocket: %s", err), http.StatusBadRequest)
-			return
-		}
-		defer wsConn.Close()
-
-		wsConn.EnableWriteCompression(true)
-		wsConn.SetCompressionLevel(flate.BestCompression)
-
-		trcConn, err := pool.Conn()
-		if err != nil {
-			http.Error(w, fmt.Sprintf("Failed to establish connection to TRC: %s", err), http.StatusInternalServerError)
-			return
-		}
-
-		ctx := r.Context()
-
-		changeCh, closeFn, err := trcConn.SubscribeStateChanges(ctx)
-		if err != nil {
-			http.Error(w, fmt.Sprintf("Failed to subscribe to state changes: %s", err), http.StatusInternalServerError)
-			return
-		}
-		defer closeFn()
-
-		var oldState *api.State
-		for {
-			select {
-			case <-ctx.Done():
-				return
-
-			case <-trcConn.Closed():
-				http.Error(w, fmt.Sprintf("Communication with TRC closed"), http.StatusServiceUnavailable)
-				return
-
-			case <-changeCh:
-				st := trcConn.State(ctx)
-				// TODO: Compute diff of st and oldState
-				_ = oldState
-				diff := st
-				oldState = st
-				if err := wsConn.WriteJSON(diff); err != nil {
-					http.Error(w, fmt.Sprintf("Failed to write state: %s", err), http.StatusInternalServerError)
-					return
-				}
-			}
-		}
-	})
-
-	http.HandleFunc("/"+commandEndpoint, func(w http.ResponseWriter, r *http.Request) {
-		defer r.Body.Close()
-
-		if r.Method != "PUT" {
-			http.Error(w, fmt.Sprintf("Expected a PUT request, got %s", r.Method), http.StatusBadRequest)
-			return
-		}
-
-		// TODO: Check token
-
-		dec := json.NewDecoder(r.Body)
-		dec.DisallowUnknownFields()
-
-		var req struct {
-			Command api.Command
-		}
-		if err := dec.Decode(&req); err != nil {
-			http.Error(w, fmt.Sprintf("Failed to read command: %s", err), http.StatusBadRequest)
-			return
-		}
-
-		trcConn, err := pool.Conn()
-		if err != nil {
-			http.Error(w, fmt.Sprintf("Failed to establish connection to TRC: %s", err), http.StatusInternalServerError)
-			return
-		}
-
-		if err := trcConn.SetCommand(r.Context(), req.Command); err != nil {
-			http.Error(w, fmt.Sprintf("Failed to send command to TRC: %s", err), http.StatusBadRequest)
-			return
-		}
-	})
-
-	http.HandleFunc("/"+turtleEndpoint+"/", func(w http.ResponseWriter, r *http.Request) {
-		defer r.Body.Close()
-
-		if r.Method != "PUT" {
-			http.Error(w, fmt.Sprintf("Expected a PUT request, got %s", r.Method), http.StatusBadRequest)
-			return
-		}
-
-		// TODO: Check token
-
-		id := strings.TrimPrefix(r.URL.Path, "/"+turtleEndpoint+"/")
-		if id == "" {
-			http.Error(w, fmt.Sprintf("An ID must be specified"), http.StatusBadRequest)
-			return
-		}
-
-		dec := json.NewDecoder(r.Body)
-		dec.DisallowUnknownFields()
-
-		var req api.TurtleState
-		if err := dec.Decode(&req); err != nil {
-			http.Error(w, fmt.Sprintf("Failed to read state: %s", err), http.StatusBadRequest)
-			return
-		}
-
-		trcConn, err := pool.Conn()
-		if err != nil {
-			http.Error(w, fmt.Sprintf("Failed to establish connection to TRC: %s", err), http.StatusInternalServerError)
-			return
-		}
-
-		if err := trcConn.SetTurtleState(r.Context(), id, &req); err != nil {
-			http.Error(w, fmt.Sprintf("Failed to send command to TRC: %s", err), http.StatusBadRequest)
-			return
-		}
-	})
-
-	http.HandleFunc("/"+turtleEndpoint, func(w http.ResponseWriter, r *http.Request) {
-		defer r.Body.Close()
-
-		if r.Method != "PUT" {
-			http.Error(w, fmt.Sprintf("Expected a PUT request, got %s", r.Method), http.StatusBadRequest)
-			return
-		}
-
-		// TODO: Check token
-
-		dec := json.NewDecoder(r.Body)
-		dec.DisallowUnknownFields()
-		defer r.Body.Close()
-
-		var req struct {
-			Turtles map[string]*api.TurtleState
-		}
-		if err := dec.Decode(&req); err != nil {
-			http.Error(w, fmt.Sprintf("Failed to read states: %s", err), http.StatusBadRequest)
-			return
-		}
-
-		trcConn, err := pool.Conn()
-		if err != nil {
-			http.Error(w, fmt.Sprintf("Failed to establish connection to TRC: %s", err), http.StatusInternalServerError)
-			return
-		}
-
-		if err := trcConn.SetState(
-			r.Context(),
-			&api.State{
-				Turtles: req.Turtles,
-			},
-		); err != nil {
-			http.Error(w, fmt.Sprintf("Failed to send command to TRC: %s", err), http.StatusBadRequest)
-			return
-		}
-	})
+	http.HandleFunc("/"+statusEndpoint, webapi.StateHandler(pool))
+	http.HandleFunc("/"+commandEndpoint, webapi.CommandHandler(pool))
+	http.HandleFunc("/"+turtleEndpoint+"/", webapi.TurtlesHandler(pool))
+	http.HandleFunc("/"+turtleEndpoint, webapi.TurtlesIDHandler(pool, func(r *http.Request) string {
+		return strings.TrimPrefix(r.URL.Path, "/"+turtleEndpoint+"/")
+	}))
 
 	if *static != "" {
 		http.Handle("/", http.FileServer(http.Dir(*static)))
