+++ pkg/trcapi/validate.go
